+++ src/main/java/bit/schedule/dto/ScheduleResponse.java
@@ -1,7 +1,6 @@
 package bit.schedule.dto;
 
 import bit.schedule.domain.Schedule;
-import bit.user.entity.UserEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -35,16 +34,6 @@ public class ScheduleResponse {
     public ScheduleResponse(Schedule schedule) {
         this.id = schedule.getId();
         this.nickName = schedule.getUser().getNickName();
-        this.title = schedule.getTitle();
-        this.content = schedule.getContent();
-        this.location = schedule.getLocation();
-        this.startDateTime = schedule.getStartDateTime();
-        this.endDateTime = schedule.getEndDateTime();
-    }
-
-    public ScheduleResponse(Schedule schedule, UserEntity user) {
-        this.id = schedule.getId();
-        this.nickName = user.getNickName();
         this.title = schedule.getTitle();
         this.content = schedule.getContent();
         this.location = schedule.getLocation();
